Исправления по результату ревью

+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/data/dao/impl/spring/AuthAuthorityDaoSpringJdbc.java
@@ -67,12 +67,6 @@ public class AuthAuthorityDaoSpringJdbc implements AuthAuthorityDao {
     @Override
     public void remove(AuthorityEntity authorityEntity) {
         JdbcTemplate jdbcTemplate = new JdbcTemplate(DataSources.dataSource(CFG.authJdbcUrl()));
-        jdbcTemplate.update(con -> {
-            PreparedStatement ps = con.prepareStatement(
-                    "DELETE FROM authority WHERE user_id = ?"
-            );
-            ps.setObject(1, authorityEntity.getId());
-            return ps;
-        });
+        jdbcTemplate.update("DELETE FROM authority WHERE user_id = ?", authorityEntity.getId());
     }
 }
